Written account-managing backend code.
