+++ src/main/java/main/Controller.java
@@ -1,12 +1,12 @@
 package main;
 
+import static main.utils.Utils.selectedItemInChoiceBox;
+import static main.utils.Utils.setChoiceBoxItems;
+
 import java.net.URL;
+import java.util.ArrayList;
 import java.util.ResourceBundle;
 
-import javafx.beans.value.ChangeListener;
-import javafx.beans.value.ObservableValue;
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
@@ -20,9 +20,18 @@ import javafx.scene.image.ImageView;
 import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.GridPane;
+import main.security.Account;
+import main.utils.FileManager;
 
 public class Controller implements Initializable {
-    Button button = null;
+
+    private final FileManager fileManager;
+
+    public Controller(FileManager fileManager) {
+        this.fileManager = fileManager;
+    }
+
+    Button lastSidebarButton = null;
 
     @FXML
     private AnchorPane homePane;
@@ -48,12 +57,26 @@ public class Controller implements Initializable {
         decryptPasswordHidden.textProperty().addListener((observable, oldValue, newValue) -> {
             decryptPasswordVisible.setText(newValue);
         });
-        decryptChoiceBox.getSelectionModel().selectedIndexProperty().addListener(new ChangeListener<Number>() {
-            @Override
-            public void changed(ObservableValue<? extends Number> observableValue, Number number, Number number2) {
-                decryptChoiceBoxSelected(observableValue, number, number2);
+        decryptChoiceBox.getSelectionModel().selectedIndexProperty().addListener(
+            (observableValue, oldIndex, newIndex) -> {
+                decryptDeleteCounter = false;
+
+                int index = newIndex.intValue();
+                if (index >= 0) {
+                    // gets the selected account
+                    Account selectedAcc = fileManager.getAccountList().get(index);
+                    String password = fileManager.getAccountPassword(selectedAcc);
+
+                    // shows the software, username and account of the selected account
+                    decryptSoftware.setText(selectedAcc.getSoftware());
+                    decryptUsername.setText(selectedAcc.getUsername());
+                    decryptPasswordVisible.setText(password);
+                    decryptPasswordHidden.setText(password);
+                } else {
+                    decryptClear();
+                }
             }
-        });
+        );
     }
 
     // #region Encrypter
@@ -81,15 +104,19 @@ public class Controller implements Initializable {
 
     @FXML
     public void encryptSave(ActionEvent event) {
+        // gets software, username and password written by the user
         String software = encryptSoftware.getText();
         String username = encryptUsername.getText();
         String password = encryptPasswordVisible.getText();
-        System.out.format("Account: %s, %s, %s\n", software, username, password);
 
-        encryptSoftware.clear();
-        encryptUsername.clear();
-        encryptPasswordVisible.clear();
-        encryptPasswordHidden.clear();
+        if (!(software.isBlank() || username.isBlank() || password.isBlank())) {
+            fileManager.addAccount(software, username, password);
+
+            encryptSoftware.clear();
+            encryptUsername.clear();
+            encryptPasswordVisible.clear();
+            encryptPasswordHidden.clear();
+        }
     }
     // #endregion
 
@@ -102,22 +129,17 @@ public class Controller implements Initializable {
 
     @FXML
     private TextField decryptSoftware, decryptUsername, decryptPasswordVisible;
-
     @FXML
     private PasswordField decryptPasswordHidden;
 
     @FXML
-    public void decryptSidebarButton(ActionEvent event) {
-        decryptSoftware.clear();
-        decryptUsername.clear();
-        decryptPasswordVisible.clear();
-        decryptPasswordHidden.clear();
+    private Button decryptDelete;
+    private boolean decryptDeleteCounter = false;
 
-        ObservableList<String> list = FXCollections.observableArrayList();
-        for (int i = 1; i < 27; i++) {
-            list.add(i + ") " + (char) (96 + i));
-        }
-        decryptChoiceBox.setItems(list);
+    @FXML
+    public void decryptSidebarButton(ActionEvent event) {
+        decryptDelete.setStyle("-fx-background-color: #a7acb1;");
+        decryptChoiceBoxLoad();
 
         decryptPane.toFront();
         setMainTitle("Decryption");
@@ -125,35 +147,80 @@ public class Controller implements Initializable {
         highlightSidebarButton(event);
     }
 
-    public void decryptChoiceBoxSelected(ObservableValue<? extends Number> observableValue, Number number, Number number2) {
-        decryptSoftware.setText("Old: " + String.valueOf(number.intValue() + 1));
-        decryptUsername.setText("New: " + String.valueOf(number2.intValue() + 1));
+    public void decryptChoiceBoxLoad() {
+        ArrayList<Account> accountList = fileManager.getAccountList();
+
+        String[] items = new String[accountList.size()];
+        StringBuilder strb;
+
+        for (int i = 0; i < items.length; i++) {
+            strb = new StringBuilder();
+
+            strb.append(i + 1)
+                    .append(") ");
+
+            Account account = accountList.get(i);
+            switch (fileManager.getLoginAccount().getSavingOrder()) {
+                case "s" -> strb.append(account.getSoftware()).append(" / ").append(account.getUsername());
+                case "u" -> strb.append(account.getUsername()).append(" / ").append(account.getSoftware());
+                default -> throw new IllegalArgumentException("Invalid language!");
+            }
+
+            items[i] = strb.toString();
+        }
+
+        setChoiceBoxItems(decryptChoiceBox, items);
+        decryptClear();
+    }
+
+    public void decryptClear() {
+        decryptSoftware.clear();
+        decryptUsername.clear();
+        decryptPasswordVisible.clear();
+        decryptPasswordHidden.clear();
+        
+        decryptChoiceBox.getSelectionModel().clearSelection();
     }
 
     @FXML
     public void decryptSave(ActionEvent event) {
-        String software = decryptSoftware.getText();
-        String username = decryptUsername.getText();
-        String password = decryptPasswordVisible.getText();
-        System.out.format("Account: %s, %s, %s\n", software, username, password);
+        int index = selectedItemInChoiceBox(decryptChoiceBox);
+        if (index >= 0) {
+            // get the new software, username and password
+            String software = decryptSoftware.getText();
+            String username = decryptUsername.getText();
+            String password = decryptPasswordVisible.getText();
 
-        decryptSoftware.clear();
-        decryptUsername.clear();
-        decryptPasswordVisible.clear();
-        decryptPasswordHidden.clear();
+            if (!(software.isBlank() && username.isBlank() && password.isBlank())) {
+                // save the new attributes of the account
+                fileManager.replaceAccount(index, software, username, password);
+
+                decryptChoiceBoxLoad();
+            }
+        }
     }
 
     @FXML
     public void decryptDelete(ActionEvent event) {
-        String software = decryptSoftware.getText();
-        String username = decryptUsername.getText();
-        String password = decryptPasswordVisible.getText();
-        System.out.format("Account: %s, %s, %s\n", software, username, password);
+        int index = selectedItemInChoiceBox(decryptChoiceBox);
+        if (index < 0) {
+            return;
+        }
 
-        decryptSoftware.clear();
-        decryptUsername.clear();
-        decryptPasswordVisible.clear();
-        decryptPasswordHidden.clear();
+        // when the deleteCounter is true it means that the user has confirmed the
+        // elimination
+        if (decryptDeleteCounter) {
+            decryptDelete.setStyle("-fx-background-color: #a7acb1;");
+
+            // removes the selected account from the list
+            fileManager.deleteAccount(index);
+
+            decryptChoiceBoxLoad();
+        } else {
+            decryptDelete.setStyle("-fx-background-color: #ff5f5f;");
+        }
+
+        decryptDeleteCounter = !decryptDeleteCounter;
     }
     // #endregion
 
@@ -179,13 +246,13 @@ public class Controller implements Initializable {
     }
 
     private void highlightSidebarButton(ActionEvent event) {
-        if (button != null) {
-            button.setStyle("-fx-background-color: #202428;");
+        if (lastSidebarButton != null) {
+            lastSidebarButton.setStyle("-fx-background-color: #202428;");
         }
 
         if (event != null) {
-            button = (Button) event.getSource();
-            button.setStyle("-fx-background-color: #42464a;");
+            lastSidebarButton = (Button) event.getSource();
+            lastSidebarButton.setStyle("-fx-background-color: #42464a;");
         }
     }
