+++ src/main/resources/fxml/styles.css
@@ -5,7 +5,7 @@
 
 .decryptButton {
     -fx-background-color: #a7acb1;
-    -fx-background-radius: 10deg;
+    -fx-background-radius: 9deg;
 }
 
 .field {
