+++ src/main/resources/old-main/Main.java
@@ -148,24 +148,6 @@ public class Main extends JFrame {
         replaceToDialogPanel(EulaAndCreditsPanel);
     }
 
-    private void formWindowClosing() {
-        // when the user shuts down the program on the first run, it won't save
-        if (loginAccount != null) {
-            try (ObjectOutputStream fOUT = new ObjectOutputStream(new FileOutputStream(filePath + "passwords.psmg"))) {
-                fOUT.writeObject(this.loginAccount);
-                fOUT.writeObject(this.accountList);
-                fOUT.close();
-
-                logger.addInfo("Files saved");
-            } catch (IOException e) {
-                logger.addError(e);
-            }
-
-            logger.addInfo("Program shut down");
-            logger.save();
-        }
-    }
-
     public void switchToProgramPanel(String @NotNull... password) {
         if (password.length > 0) {
             this.loginPassword = password[0];
@@ -201,451 +183,14 @@ public class Main extends JFrame {
         // redirects to program panel
         replaceToMainPanel(ProgramPanel);
     }
-
-    public void replaceToMainPanel(JPanel NewPanel) {
-        replacePanel(MainPanel, NewPanel);
-    }
-
-    public void replaceToDialogPanel(JPanel NewPanel) {
-        replacePanel(DialogPanel, NewPanel);
-    }
-    // #endregion
-
-    // #region Exporters
-    private void htmlMenuItemActionPerformed(ActionEvent evt) {
-        try (FileWriter file = new FileWriter(System.getProperty("user.home") + "\\Desktop\\Passwords.html")) {
-            file.write(Exporter.exportHtml(accountList, loginAccount.getLanguage(), loginPassword));
-            file.flush();
-        } catch (IOException e) {
-            logger.addError(e);
-        }
-
-        ExportAsMenu.dispatchEvent(new KeyEvent(ExportAsMenu, KeyEvent.KEY_PRESSED, 0, 0, KeyEvent.VK_ESCAPE, '←'));
-    }
-
-    private void csvMenuItemActionPerformed(ActionEvent evt) {
-        try (FileWriter file = new FileWriter(System.getProperty("user.home") + "\\Desktop\\Passwords.csv")) {
-            file.write(Exporter.exportCsv(accountList, loginPassword));
-            file.flush();
-        } catch (IOException e) {
-            logger.addError(e);
-        }
-
-        ExportAsMenu.dispatchEvent(new KeyEvent(ExportAsMenu, KeyEvent.KEY_PRESSED, 0, 0, KeyEvent.VK_ESCAPE, '←'));
-    }
     // #endregion
 
     /**
      * Redirects to the login or first run procedure, based on the password file
      * existence.
      */
-    @SuppressWarnings("unchecked")
+    
     private void run() {
-        File data_file = new File(filePath + "passwords.psmg");
-        if (data_file.getParentFile().mkdirs()) {
-            logger.addInfo("Created folder '" + data_file.getParentFile().getAbsolutePath() + "'");
-        }
-
-        // if the data file exists, it will try to read its contents
-        if (data_file.exists()) {
-            try (FileInputStream f = new FileInputStream(data_file)) {
-                ObjectInputStream fIN = new ObjectInputStream(f);
-
-                loginAccount = (LoginAccount) fIN.readObject();
-                accountList = (ArrayList<Account>) fIN.readObject();
-            } catch (IOException | ClassNotFoundException e) {
-                logger.addError(e);
-            }
-        }
-
-        if (loginAccount != null) {
-            // gets the log history
-            logger.readFile();
-
-            switch (loginAccount.getLanguage()) {
-                case "e" -> LoginPanel.load("Login", "Login");
-                case "i" -> LoginPanel.load("Accedi", "Accesso");
-                default -> throw new IllegalArgumentException("Invalid language: " + loginAccount.getLanguage());
-            }
-
-            // redirects to login panel
-            replaceToMainPanel(LoginPanel);
-        } else {
-            // redirects to first run panel
-            replaceToMainPanel(FirstRunPanel);
-        }
+        
     }
-
-    // #region AccountList methods
-    public ArrayList<Account> getAccountList() {
-        return this.accountList;
-    }
-
-    public void sortAccountList() {
-        switch (loginAccount.getSavingOrder()) {
-            case "s" -> this.accountList.sort((acc1, acc2) -> {
-                int software = acc1.getSoftware().compareTo(acc2.getSoftware());
-                return (software == 0) ? acc1.getUsername().compareTo(acc2.getUsername()) : software;
-            });
-
-            case "u" -> this.accountList.sort((acc1, acc2) -> {
-                int username = acc1.getUsername().compareTo(acc2.getUsername());
-                return (username == 0) ? acc1.getSoftware().compareTo(acc2.getSoftware()) : username;
-            });
-
-            default -> throw new IllegalArgumentException("Invalid saving order: " + loginAccount.getSavingOrder());
-        }
-    }
-    // #endregion
-
-    // #region Account methods
-    /**
-     * Sorts the account list.
-     */
-    public void addAccount(String software, String username, String password) {
-        accountList.add(Account.of(software, username, password, loginPassword));
-        sortAccountList();
-
-        logger.addInfo("Account added");
-
-        MenuBar.setVisible(true);
-    }
-
-    public void replaceAccount(int index, String software, String username, String password) {
-        if (index >= 0 && index < accountList.size()) {
-            accountList.set(index, Account.of(software, username, password, loginPassword));
-            sortAccountList();
-
-            logger.addInfo("Account edited");
-
-            MenuBar.setVisible(true);
-        }
-    }
-
-    public void deleteAccount(int index) {
-        if (index >= 0 && index < accountList.size()) {
-            accountList.remove(index);
-
-            logger.addInfo("Account deleted");
-
-            if (accountList.isEmpty()) {
-                MenuBar.setVisible(false);
-            }
-        }
-    }
-
-    public String getAccountPassword(@NotNull Account account) {
-        return account.getPassword(loginPassword);
-    }
-    // #endregion
-
-    // #region LoginAccount methods
-    public LoginAccount getLoginAccount() {
-        return this.loginAccount;
-    }
-
-    public void setLoginAccount(String savingOrder, String language, String loginPassword) {
-        this.loginAccount = LoginAccount.of(savingOrder, language, loginPassword);
-
-        if (!accountList.isEmpty()) {
-            accountList.forEach(account -> account.changeLoginPassword(this.loginPassword, loginPassword));
-            sortAccountList();
-
-            logger.addInfo("Login account changed");
-        }
-
-        this.loginPassword = loginPassword;
-    }
-    // #endregion
-
-    public Logger getLogger() {
-        return logger;
-    }
-
-    // #region Generated Code
-    private JButton DecrypterButton;
-    private DecrypterPanel DecrypterPanel;
-    private JPanel DialogPanel;
-    private JButton EncrypterButton;
-    private EncrypterPanel EncrypterPanel;
-    private JButton EulaAndCreditsButton;
-    private EulaAndCreditsPanel EulaAndCreditsPanel;
-    private JMenu ExportAsMenu;
-    private FirstRunPanel FirstRunPanel;
-    private JButton LogHistoryButton;
-    private LogHistoryPanel LogHistoryPanel;
-    private LoginPanel LoginPanel;
-    private JPanel MainPanel;
-    private JMenuBar MenuBar;
-    private JPanel ProgramPanel;
-    private JButton SettingsButton;
-    private SettingsPanel SettingsPanel;
-
-    /**
-     * This method is called from within the constructor to initialize the form.
-     * WARNING: Do NOT modify this code. The content of this method is always
-     * regenerated by the Form Editor.
-     */
-    private void initComponents() {
-        JPanel ButtonPanel = new JPanel();
-        DialogPanel = new JPanel();
-        JPanel DialogBlankPanel = new JPanel();
-        DecrypterPanel = new DecrypterPanel(this);
-        EncrypterPanel = new EncrypterPanel(this);
-        EulaAndCreditsPanel = new EulaAndCreditsPanel();
-        FirstRunPanel = new FirstRunPanel(this);
-        LogHistoryPanel = new LogHistoryPanel(this);
-        LoginPanel = new LoginPanel(this);
-        MainPanel = new JPanel();
-        ProgramPanel = new JPanel();
-        SettingsPanel = new SettingsPanel(this);
-
-        EncrypterButton = new JButton();
-        DecrypterButton = new JButton();
-        SettingsButton = new JButton();
-        LogHistoryButton = new JButton();
-        EulaAndCreditsButton = new JButton();
-
-        JLabel PasswordManagerLabel = new JLabel();
-
-        MenuBar = new JMenuBar();
-        ExportAsMenu = new JMenu();
-        JMenuItem htmlMenuItem = new JMenuItem();
-        JMenuItem csvMenuItem = new JMenuItem();
-
-        setDefaultCloseOperation(WindowConstants.EXIT_ON_CLOSE);
-        setTitle("Password Manager");
-        setBackground(new Color(0, 0, 0));
-        setCursor(new Cursor(Cursor.DEFAULT_CURSOR));
-        setFocusCycleRoot(false);
-        setFont(new Font("Dialog", Font.BOLD, 10)); // NOI18N
-        setForeground(Color.lightGray);
-        addWindowListener(new WindowAdapter() {
-            public void windowClosing(WindowEvent evt) {
-                formWindowClosing();
-            }
-        });
-
-        MainPanel.setBackground(new Color(38, 38, 38));
-        MainPanel.setForeground(new Color(242, 242, 242));
-        MainPanel.setLayout(new CardLayout());
-
-        MainPanel.add(FirstRunPanel, "card2");
-        MainPanel.add(LoginPanel, "card3");
-
-        ProgramPanel.setBackground(new Color(242, 65, 65));
-        ProgramPanel.setForeground(new Color(242, 242, 242));
-        ProgramPanel.setPreferredSize(new Dimension(1280, 720));
-
-        DialogPanel.setBackground(new Color(38, 38, 38));
-        DialogPanel.setBorder(new LineBorder(new Color(115, 41, 41), 2, true));
-        DialogPanel.setForeground(new Color(242, 242, 242));
-        DialogPanel.setLayout(new CardLayout());
-
-        DialogBlankPanel.setBackground(new Color(38, 38, 38));
-        DialogBlankPanel.setForeground(new Color(242, 242, 242));
-        DialogBlankPanel.setToolTipText("");
-
-        GroupLayout DialogBlankPanelLayout = new GroupLayout(DialogBlankPanel);
-        DialogBlankPanel.setLayout(DialogBlankPanelLayout);
-        DialogBlankPanelLayout.setHorizontalGroup(
-                DialogBlankPanelLayout.createParallelGroup(GroupLayout.Alignment.LEADING)
-                        .addGap(0, 1145, Short.MAX_VALUE));
-        DialogBlankPanelLayout.setVerticalGroup(
-                DialogBlankPanelLayout.createParallelGroup(GroupLayout.Alignment.LEADING)
-                        .addGap(0, 879, Short.MAX_VALUE));
-
-        DialogPanel.add(DialogBlankPanel, "card2");
-
-        DialogPanel.add(EncrypterPanel, "card2");
-        DialogPanel.add(DecrypterPanel, "card2");
-        DialogPanel.add(SettingsPanel, "card2");
-        DialogPanel.add(LogHistoryPanel, "card2");
-        DialogPanel.add(EulaAndCreditsPanel, "card6");
-
-        ButtonPanel.setBackground(new Color(38, 38, 38));
-        ButtonPanel.setBorder(new LineBorder(new Color(115, 41, 41), 2, true));
-        ButtonPanel.setForeground(new Color(242, 242, 242));
-        ButtonPanel.setPreferredSize(new Dimension(235, 242));
-
-        PasswordManagerLabel.setBackground(new Color(38, 38, 38));
-        PasswordManagerLabel.setFont(new Font("Dialog", Font.BOLD, 36)); // NOI18N
-        PasswordManagerLabel.setForeground(new Color(242, 242, 242));
-        PasswordManagerLabel.setHorizontalAlignment(SwingConstants.CENTER);
-        PasswordManagerLabel.setText("<html> <center> Password <br> Manager </center> </html>");
-        PasswordManagerLabel.setToolTipText("");
-        PasswordManagerLabel.setEnabled(false);
-
-        buttonSetup(EncrypterButton, this::EncrypterButtonActionPerformed);
-        buttonSetup(DecrypterButton, this::DecrypterButtonActionPerformed);
-        buttonSetup(SettingsButton, this::SettingsButtonActionPerformed);
-        buttonSetup(LogHistoryButton, this::LogHistoryButtonActionPerformed);
-        buttonSetup(EulaAndCreditsButton, this::EulaAndCreditsButtonActionPerformed);
-
-        GroupLayout ButtonPanelLayout = new GroupLayout(ButtonPanel);
-        ButtonPanel.setLayout(ButtonPanelLayout);
-        ButtonPanelLayout.setHorizontalGroup(
-                ButtonPanelLayout.createParallelGroup(GroupLayout.Alignment.LEADING)
-                        .addGroup(ButtonPanelLayout.createSequentialGroup()
-                                .addGap(6, 6, 6)
-                                .addComponent(PasswordManagerLabel, GroupLayout.DEFAULT_SIZE,
-                                        GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                                .addGap(11, 11, 11))
-                        .addGroup(ButtonPanelLayout.createSequentialGroup()
-                                .addGap(6, 6, 6)
-                                .addComponent(EncrypterButton, GroupLayout.DEFAULT_SIZE,
-                                        GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                                .addGap(11, 11, 11))
-                        .addGroup(ButtonPanelLayout.createSequentialGroup()
-                                .addGap(6, 6, 6)
-                                .addComponent(DecrypterButton, GroupLayout.DEFAULT_SIZE,
-                                        GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                                .addGap(11, 11, 11))
-                        .addGroup(ButtonPanelLayout.createSequentialGroup()
-                                .addGap(6, 6, 6)
-                                .addComponent(SettingsButton, GroupLayout.DEFAULT_SIZE,
-                                        GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                                .addGap(11, 11, 11))
-                        .addGroup(ButtonPanelLayout.createSequentialGroup()
-                                .addGap(51, 51, 51)
-                                .addComponent(LogHistoryButton, GroupLayout.DEFAULT_SIZE,
-                                        GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                                .addGap(51, 51, 51))
-                        .addGroup(ButtonPanelLayout.createSequentialGroup()
-                                .addGap(51, 51, 51)
-                                .addComponent(EulaAndCreditsButton, GroupLayout.DEFAULT_SIZE,
-                                        GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                                .addGap(51, 51, 51)));
-        ButtonPanelLayout.setVerticalGroup(
-                ButtonPanelLayout.createParallelGroup(GroupLayout.Alignment.LEADING)
-                        .addGroup(ButtonPanelLayout.createSequentialGroup()
-                                .addGap(6, 6, 6)
-                                .addComponent(PasswordManagerLabel, GroupLayout.DEFAULT_SIZE,
-                                        GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                                .addGap(6, 6, 6)
-                                .addComponent(EncrypterButton, GroupLayout.DEFAULT_SIZE,
-                                        GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                                .addGap(6, 6, 6)
-                                .addComponent(DecrypterButton, GroupLayout.DEFAULT_SIZE,
-                                        GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                                .addGap(6, 6, 6)
-                                .addComponent(SettingsButton, GroupLayout.DEFAULT_SIZE,
-                                        GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                                .addGap(6, 6, 6)
-                                .addComponent(LogHistoryButton, GroupLayout.DEFAULT_SIZE,
-                                        GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                                .addGap(6, 6, 6)
-                                .addComponent(EulaAndCreditsButton, GroupLayout.DEFAULT_SIZE,
-                                        GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                                .addGap(7, 7, 7)));
-
-        PasswordManagerLabel.getAccessibleContext().setAccessibleName("");
-
-        GroupLayout ProgramPanelLayout = new GroupLayout(ProgramPanel);
-        ProgramPanel.setLayout(ProgramPanelLayout);
-        ProgramPanelLayout.setHorizontalGroup(
-                ProgramPanelLayout.createParallelGroup(GroupLayout.Alignment.LEADING)
-                        .addGroup(ProgramPanelLayout.createSequentialGroup()
-                                .addGap(6, 6, 6)
-                                .addComponent(ButtonPanel, GroupLayout.PREFERRED_SIZE, 240,
-                                        GroupLayout.PREFERRED_SIZE)
-                                .addGap(6, 6, 6)
-                                .addComponent(DialogPanel, GroupLayout.DEFAULT_SIZE,
-                                        GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                                .addGap(5, 5, 5)));
-        ProgramPanelLayout.setVerticalGroup(
-                ProgramPanelLayout.createParallelGroup(GroupLayout.Alignment.LEADING)
-                        .addGroup(ProgramPanelLayout.createSequentialGroup()
-                                .addGap(6, 6, 6)
-                                .addGroup(ProgramPanelLayout
-                                        .createParallelGroup(GroupLayout.Alignment.LEADING)
-                                        .addComponent(ButtonPanel, GroupLayout.DEFAULT_SIZE, 240,
-                                                Short.MAX_VALUE)
-                                        .addComponent(DialogPanel, GroupLayout.DEFAULT_SIZE,
-                                                GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                                .addGap(5, 5, 5)));
-
-        MainPanel.add(ProgramPanel, "card2");
-
-        MenuBar.setBackground(new Color(38, 38, 38));
-        MenuBar.setForeground(new Color(242, 242, 242));
-
-        ExportAsMenu.setText("Export as");
-
-        htmlMenuItem.setText("HTML");
-        htmlMenuItem.addActionListener(this::htmlMenuItemActionPerformed);
-        ExportAsMenu.add(htmlMenuItem);
-
-        csvMenuItem.setText("CSV");
-        csvMenuItem.addActionListener(this::csvMenuItemActionPerformed);
-        ExportAsMenu.add(csvMenuItem);
-
-        MenuBar.add(ExportAsMenu);
-
-        setJMenuBar(MenuBar);
-
-        GroupLayout layout = new GroupLayout(getContentPane());
-        getContentPane().setLayout(layout);
-        layout.setHorizontalGroup(
-                layout.createParallelGroup(GroupLayout.Alignment.LEADING)
-                        .addComponent(MainPanel, GroupLayout.DEFAULT_SIZE,
-                                GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE));
-        layout.setVerticalGroup(
-                layout.createParallelGroup(GroupLayout.Alignment.LEADING)
-                        .addComponent(MainPanel, GroupLayout.DEFAULT_SIZE,
-                                GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE));
-
-        pack();
-    }
-
-    private static void buttonSetup(JButton Button, ActionListener ActionListener) {
-        Button.setBackground(new Color(38, 38, 38));
-        Button.setFont(new Font("Dialog", Font.BOLD, 18)); // NOI18N
-        Button.setForeground(new Color(242, 242, 242));
-        // Button.setMnemonic('?');
-        Button.addActionListener(ActionListener);
-    }
-    // #endregion
-
-    /**
-     * Main method, initializes JFrame and calls Main's constructor.
-     *
-     * @param args the command line arguments
-     */
-    public static void main(String[] args) {
-        /* Set the Nimbus look and feel */
-        // <editor-fold defaultstate="collapsed" desc=" Look and feel setting code
-        // (optional) ">
-        /*
-         * If Nimbus (introduced in Java SE 6) is not available, stay with the default
-         * look and feel.
-         * For details see
-         * http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html
-         */
-        try {
-            for (UIManager.LookAndFeelInfo info : UIManager.getInstalledLookAndFeels()) {
-                if ("Nimbus".equals(info.getName())) {
-                    UIManager.setLookAndFeel(info.getClassName());
-                    break;
-                }
-            }
-        } catch (ClassNotFoundException | InstantiationException | IllegalAccessException
-                | UnsupportedLookAndFeelException ex) {
-            java.util.logging.Logger.getLogger(Main.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        }
-        // </editor-fold>
-
-        /* Create and display the form */
-        EventQueue.invokeLater(() -> new Main().setVisible(true));
-    }
-
 }
-
-/*
- * Colors in Hex
- * #8C8A87
- * #F24141
- * #732929;
- * #F2F2F2;
- * #262626;
- */
